requirements.txt
google-cloud-storage-transfer
google-api-core